강의평가 완료

+++ src/main/java/com/green/muziuniv_be_notuser/app/shared/courseEvaluation/model/CourseEvaluationReq.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CourseEvaluationReq {
-    private int enrollmentId;
+    private Long enrollmentId;
     private String review;
     private int evScore;
 }
